pyproject.toml
[tool.poetry]
name = "chat-room"
version = "0.1.0"
description = "chat room"
authors = ["Tên Bạn <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
factory-boy = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
